gemini2.5_image_generation/pyproject.toml
[project]
name = "gemini2-5-image-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-auth>=2.40.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.28",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.9",
    "langchain-google-vertexai>=2.0.28",
    "langchain-openai>=0.3.32",
    "pillow>=11.3.0",
]
